Добавлены сложные условия из операторов && и ||, тернарный оператор, инкремент и декремент, и много чего ещё

+++ src/main/java/x590/newyava/descriptor/IncompleteMethodDescriptor.java
@@ -21,7 +21,7 @@ public record IncompleteMethodDescriptor(String name, Type returnType, @Unmodifi
 		return new IncompleteMethodDescriptor(name, returnType, Collections.unmodifiableList(arguments));
 	}
 
-	public long slots() {
+	public int slots() {
 		return MethodDescriptor.slots(arguments);
 	}
